Lots of refactoring later and keeping the existing functionality working while reducing size of controller.
