+++ stash_engine/app/controllers/stash_engine/downloads_controller.rb
@@ -3,17 +3,11 @@ require 'stash/download/file'
 require 'stash/download/version'
 # require 'rest-client'
 
-
 module StashEngine
-
-  class MerrittResponseError < StandardError
-  end
-
   class DownloadsController < ApplicationController # rubocop:disable Metrics/ClassLength
-
     before_action :setup_streaming
 
-    # set up the Merritt file & version objects so they have access to the controller context for returning responses, setting headers, etc
+    # set up the Merritt file & version objects so they have access to the controller context before continuing
     def setup_streaming
       @version_streamer = Stash::Download::Version.new(controller_context: self)
       @file_streamer = Stash::Download::File.new(controller_context: self)
@@ -24,14 +18,16 @@ module StashEngine
     def download_resource
       @resource = Resource.find(params[:resource_id])
       if @resource.may_download?(user: current_user)
-        download_full_version{ redirect_to landing_show_path(id: @resource.identifier_str, big: 'showme') }
+        @version_streamer.download(resource: @resource) do
+          redirect_to landing_show_path(id: @resource.identifier_str, big: 'showme') # if it's an async
+        end
       else
         unavailable_for_download
       end
-    rescue StashEngine::MerrittResponseError => e
+    rescue Stash::Download::MerrittResponseError => e
       # if it's a recent submission, suggest they try again later; otherwise fail
       raise e unless @resource.updated_at > Time.new - 2.hours
-      log_warning_if_needed(e)
+      Stash::Download::Base.log_warning_if_needed(e)
       # recently updated, so display a "hold your horses" message
       flash_download_unavailable
     end
@@ -64,10 +60,12 @@ module StashEngine
 
       @resource = @shares.first.resource
       unless @resource.files_public?
-        download_full_version { redirect_to private_async_form_path(id: @resource.identifier_str, big: 'showme', secret_id: params[:id]) }
-      else
-        redirect_to_public
+        @version_streamer.download(resource: @resource) do
+          redirect_to private_async_form_path(id: @resource.identifier_str, big: 'showme', secret_id: params[:id]) # for async
+          return
+        end
       end
+      redirect_to_public
     end
 
     # shows the form for private async.  Usually part of the landing page for dataset, but page may not exist for public
@@ -78,6 +76,7 @@ module StashEngine
     end
 
     def file_stream
+      file_upload = FileUpload.find(params[:file_id])
       if file_upload&.resource&.may_download?(user: current_user)
         CounterLogger.general_hit(request: request, file: file_upload)
         @file_streamer.stream_response(url: file_upload.merritt_url, tenant: current_user.tenant)
@@ -87,22 +86,6 @@ module StashEngine
     end
 
     private
-
-    def file_upload
-      @file_upload ||= FileUpload.find(params[:file_id])
-    end
-
-    # this downloads a full version as a stream from Merritt UI and takes a block with a redirect for
-    # the place to go for an asynchronous download from Merritt
-    def download_full_version
-      setup_async_download_variable
-      if @async_download
-        yield
-      else
-        CounterLogger.version_download_hit(request: request, resource: @resource)
-        @version_streamer.stream_response(url: @resource.merritt_producer_download_uri, tenant: @resource.tenant)
-      end
-    end
 
     def unavailable_for_download
       flash[:alert] = 'This dataset is private and may not be downloaded.'
@@ -125,11 +108,6 @@ module StashEngine
       Stash::Download::Version.stream_response(url: @resource.merritt_producer_download_uri, tenant: @resource.tenant)
     end
 
-    # this sets up the async download variable which it determines from Merritt
-    def setup_async_download_variable
-      @async_download = @version_streamer.merritt_async_download?(resource: @resource)
-    end
-
     def flash_download_unavailable
       flash[:notice] = [
         'This dataset was recently submitted and downloads are not yet available.',
@@ -138,11 +116,9 @@ module StashEngine
       redirect_to landing_show_path(id: @resource.identifier_str)
     end
 
-    # TODO: Move into separate class, all stuff below this (I think)
-
     # rubocop:disable Metrics/MethodLength
     def api_async_download(resource:, email:)
-      url = merritt_friendly_async_url(resource: resource)
+      url = Stash::Download::Version.merritt_friendly_async_url(resource: resource)
 
       email_from = [APP_CONFIG['contact_email']].flatten.first
       email_subject = "Your download for #{resource.title} is ready"
@@ -156,29 +132,8 @@ module StashEngine
       return if status == 200
 
       query_string = HTTP::Message.create_query_part_str(params)
-      raise_merritt_error('Merritt async download request', "unexpected status #{status}", resource.id, "#{url}?#{query_string}")
+      Stash::Download::Version.raise_merritt_error('Merritt async download request', "unexpected status #{status}", resource.id, "#{url}?#{query_string}")
     end
     # rubocop:enable Metrics/MethodLength
-
-    # TODO: move this into a merritt-specific module
-    def merritt_friendly_async_url(resource:)
-      domain, local_id = resource.merritt_protodomain_and_local_id
-      "#{domain}/asyncd/#{local_id}/#{resource.stash_version.merritt_version}"
-    end
-
-    def log_warning_if_needed(e)
-      return unless Rails.env.development?
-      msg = "MerrittResponseError checking sync/async download for resource #{@resource.id} updated at #{@resource.updated_at}"
-      backtrace = e.respond_to?(:backtrace) && e.backtrace ? e.backtrace.join("\n") : ''
-      logger.warn("#{msg}: #{e.class}: #{e}\n#{backtrace}")
-    end
-
-    def disposition_from(url)
-      "inline; filename=\"#{File.basename(URI.parse(url).path)}.zip\""
-    end
-
-    def raise_merritt_error(operation, details, resource_id, uri)
-      raise StashEngine::MerrittResponseError, "#{operation}: #{details} for resource ID #{resource_id}, URL #{uri}"
-    end
   end
 end
