+++ stash_engine/lib/stash/download/base.rb
@@ -1,6 +1,12 @@
-# Dir.glob(File.expand_path('download/*.rb', __dir__)).sort.each(&method(:require))
+require 'logger'
+
 module Stash
   module Download
+
+    class MerrittResponseError < StandardError
+    end
+
+    # this is essentially an abstract class for version and file downloads to share common methods
     class Base
       attr_reader :cc
 
@@ -19,6 +25,7 @@ module Stash
         content_type = headers.http_header['Content-Type'].try(:first)
         content_length = headers.http_header['Content-Length'].try(:first) || ''
         content_disposition = headers.http_header['Content-Disposition'].try(:first) || disposition_from(url)
+        content_disposition.gsub!(/^inline;/, 'attachment;') # do attachement if we can, not inline
         cc.response.headers['Content-Type'] = content_type if content_type
         cc.response.headers['Content-Disposition'] = content_disposition
         cc.response.headers['Content-Length'] = content_length
@@ -26,6 +33,18 @@ module Stash
         cc.response_body = Stash::Streamer.new(client, url)
       end
       # rubocop:enable Metrics/AbcSize
+
+      def disposition_from(url)
+        "inline; filename=\"#{File.basename(URI.parse(url).path)}.zip\""
+      end
+
+      def self.log_warning_if_needed(e)
+        return unless Rails.env.development?
+        msg = "MerrittResponseError checking sync/async download for resource #{@resource.id} updated at #{@resource.updated_at}"
+        backtrace = e.respond_to?(:backtrace) && e.backtrace ? e.backtrace.join("\n") : ''
+        Rails.logger.warn("#{msg}: #{e.class}: #{e}\n#{backtrace}")
+      end
+
     end
   end
 end
